feat: contracts deployed and clenup

+++ contracts/ERC20Registry.sol
@@ -7,7 +7,6 @@ import "@openzeppelin/contracts/access/Ownable.sol";
 /// @author Paul Czajka [paul.czajka@gmail.com]
 /// @notice Tokenscope ERC20 token registry
 contract ERC20Registry is Ownable {
-
     // This contract stores up to 256 unique facts for each token.
     // A fact has two parts:
     //   id: a uint8 value, 0 - 255
@@ -37,9 +36,8 @@ contract ERC20Registry is Ownable {
     // with a uint256-casted fact id.  Use factsToFactSet() to convert individual
     // fact ids to a correct factSet representation.
 
-
     // Standard facts
-    uint8 public constant IS_REGISTERED = 0;  // Token exists in this contract. Set to 1 when added to registry
+    uint8 public constant IS_REGISTERED = 0; // Token exists in this contract. Set to 1 when added to registry
     uint8 public constant IS_VALID_ERC20 = 1; // Token conforms to ERC20 standard. Set by governance
 
     /// High-water mark for the highest fact id (factSet bit-position)
@@ -58,11 +56,10 @@ contract ERC20Registry is Ownable {
     /// @param validatedFacts The new validated fact set of the token
     event ERC20ValidatedFacts(address indexed token, uint256 validatedFacts);
 
-
     /// @param _governor The owning Governance contract
     constructor(address _governor) {
         // Governance contract is owner
-        transferOwnership(_governor); 
+        transferOwnership(_governor);
 
         // factCreated events are the effective "catalog" of facts available.
         // Emit the first to common facts so they show up in the catalog like all the rest.
@@ -73,7 +70,10 @@ contract ERC20Registry is Ownable {
     /// The token must be registered
     /// @param _token The token
     modifier isRegistered(address _token) {
-        require(tokenFacts[_token] & IS_REGISTERED == IS_REGISTERED, "TOKEN_NOT_REGISTERED");
+        require(
+            tokenFacts[_token] & IS_REGISTERED == IS_REGISTERED,
+            "TOKEN_NOT_REGISTERED"
+        );
         _;
     }
 
@@ -84,17 +84,23 @@ contract ERC20Registry is Ownable {
             // The highest valid factSet value would have all bits set to 1 for all created facts.
             // "1 << (highwaterFact + 1)" creates a factSet that exceeds this number by 1:
             // subtracting one yields the value where all bits are set to 1 for all created facts.
-            require(_factSet <= (1 << (highwaterFact + 1)) - 1, "INVALID_FACT_SET");
+            require(
+                _factSet <= (1 << (highwaterFact + 1)) - 1,
+                "INVALID_FACT_SET"
+            );
         }
         _;
     }
 
-    function _isValidated(address _token, uint256 _factSet) private view returns (bool) {
+    function _isValidated(address _token, uint256 _factSet)
+        private
+        view
+        returns (bool)
+    {
         // An individual fact is validated if its bit-position is set to 1.
         // We can validate multiple facts at once.
         return tokenFacts[_token] & _factSet == _factSet;
     }
-
 
     /*/////////////////////////////////////////////////////////////////////////////////
         Registry Administration
@@ -114,14 +120,17 @@ contract ERC20Registry is Ownable {
     /// @param _token The token
     /// @param _factSet The set of all validated facts for the token
     /// @dev If the token already exists, its present factSet will be entirely overwritten by this new value.
-    function addUpdateERC20(address _token, uint256 _factSet) external onlyOwner validFactSet(_factSet) {
+    function addUpdateERC20(address _token, uint256 _factSet)
+        external
+        onlyOwner
+        validFactSet(_factSet)
+    {
         // The IS_REGISTERED attr is always true in storage
         //  (1 << IS_REGISTERED) = 1
         tokenFacts[_token] = _factSet | 1;
 
         emit ERC20ValidatedFacts(_token, _factSet | 1);
     }
-
 
     /*/////////////////////////////////////////////////////////////////////////////////
         Registry Querying
@@ -138,7 +147,12 @@ contract ERC20Registry is Ownable {
     /// Convenience method to determine if a particular token is a valid ERC20 implementation
     /// @param _token The token
     /// @return bool
-    function tokenIsValidERC20(address _token) external view isRegistered(_token) returns (bool) {
+    function tokenIsValidERC20(address _token)
+        external
+        view
+        isRegistered(_token)
+        returns (bool)
+    {
         // Second argument: 1 << IS_VALID_ERC20 = 2
         return _isValidated(_token, 2);
     }
@@ -147,7 +161,12 @@ contract ERC20Registry is Ownable {
     /// @param _token The token
     /// @param _facts The array of uint8 fact ids to be validated
     /// @return bool
-    function factsAreValidated(address _token, uint8[] calldata _facts) external view isRegistered(_token) returns (bool) {
+    function factsAreValidated(address _token, uint8[] calldata _facts)
+        external
+        view
+        isRegistered(_token)
+        returns (bool)
+    {
         return factSetIsValidated(_token, factsToFactSet(_facts));
     }
 
@@ -156,7 +175,13 @@ contract ERC20Registry is Ownable {
     /// any facts above and beyond the flagged ones are not accounted for and wil have no impact on the result.
     /// @param _token The token
     /// @param _factSet The flattened uint256 fact set to be validated
-    function factSetIsValidated(address _token, uint256 _factSet) public view isRegistered(_token) validFactSet(_factSet) returns (bool) {
+    function factSetIsValidated(address _token, uint256 _factSet)
+        public
+        view
+        isRegistered(_token)
+        validFactSet(_factSet)
+        returns (bool)
+    {
         return _isValidated(_token, _factSet);
     }
 
@@ -168,9 +193,13 @@ contract ERC20Registry is Ownable {
     /// @dev Does not validate that any particular fact values exist
     /// @param _facts The array of fact values to convert into a fact set
     /// @return factSet
-    function factsToFactSet(uint8[] calldata _facts) public pure returns (uint256 factSet) {
-        uint len = _facts.length;
-        for (uint i = 0 ; i < len; ++i) {
+    function factsToFactSet(uint8[] calldata _facts)
+        public
+        pure
+        returns (uint256 factSet)
+    {
+        uint256 len = _facts.length;
+        for (uint256 i = 0; i < len; ++i) {
             factSet = factSet | (1 << _facts[i]);
         }
     }
@@ -179,8 +208,11 @@ contract ERC20Registry is Ownable {
     /// @dev Does not validate that any particular fact values exist
     /// @param _factSet The fact set to convert into an array of fact values
     /// @return uint8[]
-    function factSetToFacts(uint256 _factSet) external pure returns (uint8[] memory) {
-
+    function factSetToFacts(uint256 _factSet)
+        external
+        pure
+        returns (uint8[] memory)
+    {
         // We can't create a dynamic memory array, so we need to loop twice:
         // 1) Discover the number of facts. Then we size the facts array appropriately
         // 2) Populate the facts array
